requirements.txt
Django==5.2.7
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
